+++ R/package.R
@@ -16,7 +16,7 @@ NULL
 # working with GSEA results.
 utils::globalVariables(
   c(".",
-    # appears in data.table[manip := ulations] of core multiGSEA tables
+    # appears in data.table[manip := ulations] of core sparrow tables
     "active", "collection", "feature_id", "N", "n", "name",
     "mean.logFC.trim",
     "pval", "padj", "padj.by.collection", "padj.up", "padj.down",