+++ R/zzz.R
@@ -1,19 +1,12 @@
+
+# This looks like the way tiledbr packages does it, thanks Dirk!
+.pkgcache <- new.env(parent = emptyenv())
+
 .onLoad <- function(libname, pkgname) {
   ## Setup default option values
   opts <- options()
 
-  # pkg.opts <- list(
-  #   multiGSEA.df.return='data.frame')
-  # toset <- !(names(pkg.opts) %in% names(opts))
-  # if (any(toset)) {
-  #   options(pkg.opts[toset])
-  # }
-  #
-  # df.return <- getOption('multiGSEA.df.return')
-  # if (!df.return %in% c('data.table', 'data.frame')) {
-  #   warning("Invalid value for options(multiGSEA.df.return) (",
-  #           df.return, "), setting to 'data.table'", immediate.=TRUE)
-  # }
-
+  species <- msigdbr::msigdbr_species()[["species_name"]]
+  .pkgcache[["msigdb"]] <- sapply(species, function(x) NULL, simplify = FALSE)
   invisible()
 }