+++ tests/testthat/setup.R
@@ -1,7 +1,11 @@
-# library("magrittr")
-# library("data.table")
-# library("dtplyr")
-# library("dplyr")
+# testthat dogma no longer suggests we use helper-* files, just setup.R
+# Note that helper-*.R files are loaded on devtools::load_all() by default
+# but the setup*.R files are not
+
+library("magrittr")
+library("data.table")
+library("dtplyr")
+library("dplyr")
 
 # Since loading data from msigdbr can be a bit slow, we load a serious of
 # GeneSetDb objects here that will be used throughout the test harness.